backend/ingest/requirements.txt
beautifulsoup4
httpx
pydantic-settings
sqlalchemy[asyncio]
asyncpg
gspread
google-auth
